Show value for unknown OID in DN
In case the OID is not known, but the value is a printable string, show
the value in the DN. Handling unknown OIDs in general is hard, as they
can contain arbitrary data. However, a common case is that that the value
is just a printable string, which can easily be represented in the DN.

If you are interested, I would also be happy to implement a mechanism
to register encoders for unknown OIDs.